requirements.txt

playwright==1.44.0
aiohttp==3.9.5
Pillow==10.2.0
scikit-image==0.22.0
